+++ util/opentelemetry-util-http/tests/test_remove_credentials.py
@@ -10,22 +10,28 @@ class TestRemoveUrlCredentials(unittest.TestCase):
         self.assertEqual(cleaned_url, url)
 
     def test_remove_credentials(self):
-        url = "http://someuser:somepass@opentelemetry.io:8080/test/path?query=value"
+        url = "http://someuser:somepass@opentelemetry.io:8080/test/path?sig=value"
         cleaned_url = remove_url_credentials(url)
         self.assertEqual(
-            cleaned_url, "http://opentelemetry.io:8080/test/path?query=value"
+            cleaned_url, "http://REDACTED:REDACTED@opentelemetry.io:8080/test/path?sig=value"
         )
 
     def test_remove_credentials_ipv4_literal(self):
         url = "http://someuser:somepass@127.0.0.1:8080/test/path?query=value"
         cleaned_url = remove_url_credentials(url)
         self.assertEqual(
-            cleaned_url, "http://127.0.0.1:8080/test/path?query=value"
+            cleaned_url, "http://REDACTED:REDACTED@127.0.0.1:8080/test/path?query=value"
         )
 
     def test_remove_credentials_ipv6_literal(self):
         url = "http://someuser:somepass@[::1]:8080/test/path?query=value"
         cleaned_url = remove_url_credentials(url)
         self.assertEqual(
-            cleaned_url, "http://[::1]:8080/test/path?query=value"
+            cleaned_url, "http://REDACTED:REDACTED@[::1]:8080/test/path?query=value"
         )
+
+    def test_empty_url(self):
+        url = ""
+        cleaned_url = remove_url_credentials(url)
+        self.assertEqual(cleaned_url, url)
+