+++ CHANGELOG.md
@@ -73,7 +73,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `opentelemetry-instrumentation-aiohttp-client` Add support for HTTP metrics
-  ([#3517](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3517))
+  ([#3517](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3517))  
+- `opentelemetry-util-http` Added support for redacting specific url query string values and url credentials in instrumentations
+  ([#3508](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3508))
 - `opentelemetry-instrumentation-httpx` Add support for HTTP metrics
   ([#3513](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3513))
 - `opentelemetry-instrumentation` Allow re-raising exception when instrumentation fails